+++ lib/src/navbar_router.dart
@@ -323,6 +323,15 @@ class _NavbarRouterState extends State<NavbarRouter>
   }
 }
 
+/// `navigate` method is deprecated use Navigate.push instead, see [routes-and-navigation](https://docs.page/maheshmnj/navbar_router/routes-and-navigation) for details.
+@Deprecated('Use `Navigate.push` instead')
+Future<void> navigate(BuildContext context, String route,
+        {bool isDialog = false,
+        bool isRootNavigator = true,
+        Map<String, dynamic>? arguments}) =>
+    Navigator.of(context, rootNavigator: isRootNavigator)
+        .pushNamed(route, arguments: arguments);
+
 final NavbarNotifier _navbarNotifier = NavbarNotifier();
 List<Color> colors = [mediumPurple, Colors.orange, Colors.teal];
 const Color mediumPurple = Color.fromRGBO(79, 0, 241, 1.0);