+++ test/navbar_utils.dart
@@ -185,7 +185,6 @@ class _ProductListState extends State<ProductList> {
                   onTap: () {
                     NavbarNotifier.hideBottomNavBar = false;
                     Navigate.pushNamed(context, ProductDetail.route,
-                        isRootNavigator: false,
                         arguments: {'id': index.toString()});
                   },
                   child: ProductTile(index: index)),
@@ -246,7 +245,6 @@ class ProductDetail extends StatelessWidget {
               onPressed: () {
                 NavbarNotifier.hideBottomNavBar = false;
                 Navigate.pushNamed(context, ProductComments.route,
-                    isRootNavigator: false,
                     arguments: {'id': argsId.toString()});
               },
               child: const Text('show comments'))
@@ -296,8 +294,7 @@ class UserProfile extends StatelessWidget {
             IconButton(
               icon: const Icon(Icons.edit),
               onPressed: () {
-                Navigate.pushNamed(context, ProfileEdit.route,
-                    isRootNavigator: false);
+                Navigate.pushNamed(context, ProfileEdit.route);
               },
             )
           ],